buffer bugfix: changed dtype to input elems.dtype

+++ src/buffers/augmented_representations_buffer.py
@@ -39,7 +39,7 @@ class AugmentedRepresentationsBuffer:
             # Extend buffer to have same dim of batch_x
             buffer_shape = list(batch_x.size())
             buffer_shape[0] = 0
-            self.buffer = torch.empty(buffer_shape).to(self.device)
+            self.buffer = torch.empty(buffer_shape, dtype=batch_x.dtype).to(self.device)
 
             # # Extend buffer_features to have same dim of batch_features
             # buffer_shape = list(batch_features.size())
